New agent mode: restore prometheus
